* sdcc.sln, * src/yacc.vcxproj: deleted yacc project * src/sdcc.vcxproj, * src/sdcc.vcxproj.filters: moved SDCC.y here for better dependencies * SDCC.props: new, added Visual C property sheet for generic settings like   boost path (default: c:\local\boost_1_59_0\) * src/hc08/hc08.vcxproj, * src/sdcc.vcxproj, * src/stm8/stm8.vcxproj, * src/z80/z80.vcxproj: use SDCC.props * sdas/linksrc/lkaomf51.c: output capitals depending on the zflag * src/common.h * src/stm8/ralloc2.cc (extra_ic_generated), * src/z80/ralloc2.cc (default_operand_cost, assign_cost, Ainst_ok,   rough_cost_estimate), * src/SDCCglue.c (printIvalChar), * src/SDCCopt.c (convertToFcall): fixed warnings * src/SDCCutil.c (isblank): added for MSVC 2010 * support/cpp/ansidecl.h: deleted * support/cpp/sdcpp.vcxproj, * support/cpp/sdcpp.vcxproj.filters: use files in support/sdbinutils/ * support/sdbinutils/libiberty/getpwd.c: fixed for MSVC
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@9384 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCglue.c
@@ -1024,9 +1024,9 @@ int
 printIvalChar (symbol * sym, sym_link * type, initList * ilist, struct dbuf_s *oBuf, const char *s, bool check)
 {
   value *val;
-  unsigned int size = DCL_ELEM (type);
+  size_t size = DCL_ELEM(type);
   char *p;
-  long asz;
+  size_t asz;
 
   if (!s)
     {
